+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,5 @@
 blank_issues_enabled: false
 contact_links:
   - name: Community Support
-    url: https://github.com/Xmaster6y/xrl/discussions
+    url: https://github.com/Xmaster6y/xdrl/discussions
     about: Please ask and answer questions here.
