+++ docs/source/_static/images/xdrl-logo.png
