+++ docs/source/_static/switcher.json
@@ -1,6 +1,6 @@
 [
     {
       "version": "dev",
-      "url": "https://xrl.readthedocs.io/en/latest/"
+      "url": "https://xdrl.readthedocs.io/en/latest/"
     }
   ]
