+++ src/xdrl/__init__.py
@@ -1,6 +1,6 @@
 from importlib.metadata import PackageNotFoundError, version
 
 try:
-    __version__ = version("xrl")
+    __version__ = version("xdrl")
 except PackageNotFoundError:
     __version__ = "unknown version"
